Extract update-state function
