+++ frontend/src/app/admin/programs/program-overview/helpers/index.ts
@@ -1,3 +1,4 @@
 export * from './program-state';
 export * from './filter-message';
+export * from './update-state';
 export * from './program';
