+++ frontend/src/app/admin/programs/program-overview/helpers/update-state.ts
@@ -1,0 +1,44 @@
+import { ApplicationFacingProgram } from '../../../models/program';
+import { FilterMessage } from './filter-message';
+import { ProgramState } from './program-state';
+import { Observable } from 'rxjs';
+import { scan } from 'rxjs/operators'
+
+export function updateState(input$: Observable<FilterMessage | ApplicationFacingProgram[] | ApplicationFacingProgram>) : Observable<ProgramState> {
+  
+  const INITIAL_STATE = new ProgramState([], new FilterMessage({ type: '', value: 'none' }));
+  
+  return input$
+    .pipe(
+      scan((state: ProgramState, update: FilterMessage | ApplicationFacingProgram[] | ApplicationFacingProgram) => {
+        if (update instanceof FilterMessage) {
+          state.filter = update;
+          return state;
+        } else if (Array.isArray(update)) {
+          state.programs = [...update].filter(program => program.guid !== undefined && program.guid !== null).sort(programComparator);
+          return state;
+        } else if (typeof update === 'object' && update.guid !== undefined) {
+          const index = state.programs.findIndex(p => p.guid === update.guid)
+    
+          if (index >= 0) {
+            state.programs[index] = update;
+          }
+          return state;
+        }
+    
+        return INITIAL_STATE;
+      }, INITIAL_STATE)
+    )
+}
+
+
+function programComparator(a: ApplicationFacingProgram, b: ApplicationFacingProgram): number {
+  const titleA = a.user.title.toUpperCase();
+  const titleB = b.user.title.toUpperCase();
+
+  if (titleA < titleB) return -1;
+
+  if (titleB < titleA) return 1;
+
+  return 0;
+}
